Finish filter

+++ c_test/test.cpp
@@ -226,7 +226,7 @@ int main() {
 
     std::cout << ms_double.count() << "ms\n";
 
-    for (int i=0; i<rects.size(); i++) {
+    for (size_t i=0; i<rects.size(); i++) {
         cout << rects[i].center() << " " << rects[i].empty() << endl;
     }
